refactor(core): extract shared predicate DSL base from workflows
Move grammar, evaluation semantics, and label rendering to
core/src/predicate; workflows/predicate becomes the workflow-typed
target composed via createPredicateEvaluator. Public
@mastra/core/workflows surface unchanged via re-exports. Adds
definition-time root validation helper and pins root-presence
semantics with a test.
